% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkModule.R
\name{linkModServer}
\alias{linkModServer}
\title{Server-side of clickable drop down list module}
\usage{
linkModServer(id)
}
\arguments{
\item{id}{character}
}
\description{
Server-side of clickable drop down list module
}
