% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExamples.R
\name{seeExampleCode}
\alias{seeExampleCode}
\title{See the code for an example shiny apps}
\usage{
seeExampleCode(example)
}
\arguments{
\item{example}{either "simple" or "bcstatsheader"}
}
\description{
See the code for an example shiny apps
}
