% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkModule.R
\name{linkModServer}
\alias{linkModServer}
\title{Server-side of clickable drop down list module}
\usage{
linkModServer(id, data_loader = load_link_csv)
}
\arguments{
\item{id}{character}

\item{data_loader}{optional function that returns a data.frame with name and url columns}
}
\description{
Server-side of clickable drop down list module
}
