% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkModule.R
\name{linkModUI}
\alias{linkModUI}
\title{Module UI for clickable drop down list with links to BC Stats apps}
\usage{
linkModUI(id)
}
\arguments{
\item{id}{character}
}
\value{
a UI element
}
\description{
For examples run \code{simpleModuleDemo()} or \code{bcstatsHeaderDemo()}.
To see the underlying code for these examples run \code{simpleModuleDemoCode()}
or \code{bcstatsHeaderDemoCode()}.
}
